add session helper and test wip
