+++ app/helpers/session_helper.rb
@@ -1,0 +1,9 @@
+module SessionHelper
+  def clear_session
+    user_session['idv/doc_auth'] = {}
+    user_session['idv/in_person'] = {}
+    user_session['idv/inherited_proofing'] = {}
+    idv_session.clear
+    Pii::Cacher.new(current_user, user_session).delete
+  end
+end
